requirements.txt
fastapi
uvicorn
supabase
sqlalchemy
aiosqlite
pydantic-settings